Reduce copy overhead with generational allocator.
Currently when we clone a minisat solver we create a full copy of the
clause allocator's memory. Since we mainly copy solvers for creation
of checkpoints the solver's will share a lot of clauses and clauses are
never modified.

To reduce overhead we split clause allocations into generations with
a new generation being created after each copy operation. All previous
generations will be shared between the solvers. To ensure that memory
is still freed correctly, generations are reference counted and deleted
once there are no references held.